+++ samples/qa-a11y/src/views/Scroller.js
@@ -13,9 +13,9 @@ class ScrollerView extends Component {
 		};
 	}
 
-	onClickChangeAriaLabelButton = () => this.setState((state) => ({customAriaLabel: !state.customAriaLabel}));
+	onToggleChangeAriaLabelButton = () => this.setState((state) => ({customAriaLabel: !state.customAriaLabel}));
 
-	onClickChangeJSNativeButton = () => this.setState((state) => ({isNative: !state.isNative}));
+	onToggleChangeJSNativeButton = () => this.setState((state) => ({isNative: !state.isNative}));
 
 	render () {
 		const
@@ -27,14 +27,14 @@ class ScrollerView extends Component {
 				<Cell shrink>
 					<ToggleButton
 						size="small"
-						onClick={this.onClickChangeAriaLabelButton}
+						onToggle={this.onToggleChangeAriaLabelButton}
 						selected={customAriaLabel}
 					>
 						Customizable aria-labels on ScrollButtons
 					</ToggleButton>
 					<ToggleButton
 						size="small"
-						onClick={this.onClickChangeJSNativeButton}
+						onToggle={this.onToggleChangeJSNativeButton}
 						selected={isNative}
 					>
 						Native