+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [4.0.4] - 2021-12-29
+
+### Fixed
+
+- `moonstone/TooltipDecorator` arrow position
+
+## [4.0.3] - 2021-09-29
+
+### Fixed
+
+- `samples/sampler` not to fail in sampler build
+
+
 ## [4.0.2] - 2021-07-29
 
 ### Fixed